+++ marimo/__init__.py
@@ -135,7 +135,7 @@ from marimo._runtime.capture import (
 )
 from marimo._runtime.context.utils import running_in_notebook
 from marimo._runtime.control_flow import MarimoStopError, stop
-from marimo._runtime.runtime import (
+from marimo._runtime.runtime.directives import (
     app_meta,
     cli_args,
     defs,
