+++ marimo/_runtime/context/kernel_context.py
@@ -25,7 +25,7 @@ if TYPE_CHECKING:
 
     from marimo._ast.app import InternalApp
     from marimo._messaging.types import Stream
-    from marimo._runtime.runtime import Kernel
+    from marimo._runtime.runtime.kernel import Kernel
     from marimo._runtime.state import State
     from marimo._types.ids import CellId_t
